pyproject.toml
[project]
name = "ecommerce-image-validator"
version = "0.1.0"
description = "ML-powered e-commerce product image quality validator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "ultralytics>=8.0.0",
    "langchain-groq>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
bonus = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "anthropic>=0.7.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
